Move url define to outside of SDK
